requirements.txt
fastapi
uvicorn
streamlit
opencv-python
numpy
pillow
pdfplumber
scikit-learn
